move babel.config.js into build-settings
