+++ build-settings/babel.config.js
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-commonjs */
 module.exports = {
     presets: ["@babel/preset-flow", "@babel/preset-react", "@babel/preset-env"],
     plugins: [
