+++ build-settings/rollup.config.js
@@ -3,7 +3,7 @@ import fs from "fs";
 import autoExternal from "rollup-plugin-auto-external";
 import babel from "rollup-plugin-babel";
 
-const {presets, plugins} = require("../babel.config.js");
+const {presets, plugins} = require("./babel.config.js");
 
 const createConfig = (pkgName) => ({
     output: {
