--- gen-webpack-config.js
@@ -1,38 +1,0 @@
-const path = require("path");
-
-module.exports = function(subPkgRoot) {
-    const subPkgJson = require(path.join(subPkgRoot, "./package.json"));
-    const subPkgDeps = subPkgJson.dependencies
-        ? Object.keys(subPkgJson.dependencies)
-        : [];
-
-    const rootPkgJson = require(path.join(subPkgRoot, "../../package.json"));
-    const rootPkgDeps = rootPkgJson.dependencies
-        ? Object.keys(rootPkgJson.dependencies)
-        : [];
-
-    return {
-        entry: "./index.js",
-        output: {
-            libraryTarget: "commonjs2",
-            filename: "dist/index.js",
-            path: subPkgRoot,
-        },
-        externals: [...rootPkgDeps, ...subPkgDeps],
-        module: {
-            rules: [
-                {
-                    test: /\.js$/,
-                    exclude: /node_modules/,
-                    use: {
-                        loader: "babel-loader",
-                    },
-                },
-            ],
-        },
-        optimization: {
-            minimize: false,
-        },
-        node: {process: false},
-    };
-};
