+++ utils/jest.transform.js
@@ -1,5 +1,5 @@
 const babelJest = require("babel-jest");
 
-const babelConfig = require("../babel.config.js");
+const babelConfig = require("../build-settings/babel.config.js");
 
 module.exports = babelJest.createTransformer(babelConfig);
